Rm the binary_search from SliceEvents setup
Specifically, it appears that despite correct Ord, even if a
binary_search finds a valid element, there may be another intersecting
span preceding the range produced by range_intersect separated by
another non-intersecting event.

This can happen when an earlier element with a very long span does
intersect and it precedes an element with a short span that does not
intersect, followed by another span that does intersect.

This is because events are primarily ordered by their start before
they are ordered by their end.

To solve this we'll need to either:

- manually separate out valid elements that fall within the start cull
  range. This doesn't solve SliceEvents behaving incorrecly in general
  though...
- Abandon binary search. It's effectiveness is ultimately based on
  heuristic and isn't consistent. It means that there can be any
  number of intersecting events falling before the discovered start.

Abandoning binary search approach (for the reasons mentioned above) in
favour of a simple retain in `EventCache::update`. `SliceEvents` now
iters all events, returning Some if there's an intersection. Neither
of these are as efficient, but they're correct.